feat: create current color
