+++ src/main/scala/ulisse/infrastructures/view/components/decorators/SwingEnhancements.scala
@@ -1,7 +1,7 @@
 package ulisse.infrastructures.view.components.decorators
 
 import ulisse.infrastructures.view.common.Observers
-import ulisse.infrastructures.view.components.styles.Styles
+import ulisse.infrastructures.view.components.styles.{CurrentColor, Styles}
 import ulisse.infrastructures.view.components.styles.Styles.EnhancedLookExtensions.*
 import ulisse.infrastructures.view.components.styles.Styles.Palette
 
@@ -12,24 +12,6 @@ import scala.swing.event.MouseEvent
 
 @SuppressWarnings(Array("org.wartremover.warts.Var"))
 object SwingEnhancements:
-
-  /** Represent the current color of the component. */
-  final case class CurrentColor(private var _current: Color):
-    def current: Color                                    = _current
-    def current_=(color: Color): Unit                     = _current = color
-    private def hoverColor(palette: Styles.Palette): Unit = palette.hoverColor.foreach(current = _)
-    private def clickColor(palette: Styles.Palette): Unit = palette.clickColor.foreach(current = _)
-    private def exitColor(palette: Styles.Palette): Unit =
-      palette.hoverColor.foreach(_ => current = palette.background)
-    private def releaseColor(palette: Styles.Palette): Unit =
-      palette.clickColor.foreach(_ => current = palette.background)
-
-    /** Initialize the color reactions of the component. */
-    def initColorReactions(component: EnhancedLook, palette: () => Palette): Reactions.Reaction =
-      case _: event.MousePressed  => clickColor(palette()); component.updateGraphics()
-      case _: event.MouseReleased => releaseColor(palette()); component.updateGraphics()
-      case _: event.MouseEntered  => hoverColor(palette()); component.updateGraphics()
-      case _: event.MouseExited   => exitColor(palette()); component.updateGraphics()
 
   /** Base trait decorator to enhanced look of swing [[Component]] */
   trait EnhancedLook extends Component:
